Get is_type_supported out of MediaSource
