+++ gecko-media/src/mse/mediasource.rs
@@ -2,10 +2,8 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-use bindings::{GeckoMedia_MediaSource_IsTypeSupported, GeckoMediaSourceImpl};
-use std::ffi::CString;
+use bindings::GeckoMediaSourceImpl;
 use std::rc::Rc;
-use std::sync::mpsc;
 
 def_gecko_media_struct!(MediaSource);
 
@@ -13,21 +11,7 @@ impl_new_gecko_media_struct!(MediaSource, MediaSourceImpl, GeckoMedia_MediaSourc
 
 impl_drop_gecko_media_struct!(MediaSource, GeckoMedia_MediaSource_Shutdown);
 
-impl MediaSource {
-    pub fn is_type_supported(&self, mime_type: &str) -> bool {
-        if let Ok(mime_type) = CString::new(mime_type.as_bytes()) {
-            let (sender, receiver) = mpsc::channel();
-            self.gecko_media.queue_task(move || unsafe {
-                sender
-                    .send(GeckoMedia_MediaSource_IsTypeSupported(mime_type.as_ptr()))
-                    .unwrap();
-            });
-            receiver.recv().unwrap()
-        } else {
-            false
-        }
-    }
-}
+impl MediaSource {}
 
 /// Users of MediaSource pass in an implementation of this trait when creating
 /// MediaSource objects.
